게시판 soft delete 제거, hard delete 채택
