+++ SpringBoot/src/main/java/com/seungwook/ktsp/domain/board/function/etc/notice/controller/NoticeCommandController.java
@@ -2,7 +2,7 @@ package com.seungwook.ktsp.domain.board.function.etc.notice.controller;
 
 import com.seungwook.ktsp.domain.board.function.etc.common.dto.request.BoardRegisterRequest;
 import com.seungwook.ktsp.domain.board.function.etc.common.dto.request.BoardUpdateRequest;
-import com.seungwook.ktsp.domain.board.function.etc.notice.service.NoticeCommandService;
+import com.seungwook.ktsp.domain.board.function.etc.notice.service.NoticeService;
 import com.seungwook.ktsp.global.auth.support.AuthHandler;
 import com.seungwook.ktsp.global.response.Response;
 import jakarta.validation.Valid;
@@ -15,12 +15,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/admin/notice")
 public class NoticeCommandController {
 
-    private final NoticeCommandService noticeCommandService;
+    private final NoticeService noticeService;
 
     @PostMapping
     public ResponseEntity<Response<Void>> registerNotice(@Valid @RequestBody BoardRegisterRequest request) {
 
-        noticeCommandService.registerNotice(AuthHandler.getUserId(), request);
+        noticeService.registerNotice(AuthHandler.getUserId(), request);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("공지사항 등록 성공")
@@ -30,7 +30,7 @@ public class NoticeCommandController {
     @PatchMapping("/{boardId}")
     public ResponseEntity<Response<Void>> updateNotice(@PathVariable long boardId, @Valid @RequestBody BoardUpdateRequest request) {
 
-        noticeCommandService.updateNotice(boardId, request);
+        noticeService.updateNotice(boardId, request);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("공지사항 수정 성공")
@@ -40,7 +40,7 @@ public class NoticeCommandController {
     @DeleteMapping("/{boardId}")
     public ResponseEntity<Response<Void>> deleteNotice(@PathVariable long boardId) {
 
-        noticeCommandService.deleteNotice(boardId);
+        noticeService.deleteNotice(boardId);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("공지사항 삭제 성공")
