--- SpringBoot/src/main/java/com/seungwook/ktsp/domain/board/function/etc/notice/service/NoticeQueryService.java
@@ -1,19 +1,0 @@
-package com.seungwook.ktsp.domain.board.function.etc.notice.service;
-
-import com.seungwook.ktsp.domain.board.function.etc.notice.entity.Notice;
-import com.seungwook.ktsp.domain.board.function.etc.notice.service.domain.NoticeDomainService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class NoticeQueryService {
-
-    private final NoticeDomainService noticeDomainService;
-
-    @Transactional(readOnly = true)
-    public Notice getNotice(long boardId) {
-        return noticeDomainService.findAsNotice(boardId);
-    }
-}
