+++ SpringBoot/src/main/java/com/seungwook/ktsp/domain/file/service/FileService.java
@@ -31,15 +31,16 @@ public class FileService {
     @Transactional
     public String uploadFile(MultipartFile file, boolean isImageFile) {
 
+        // 파일 저장
         UploadFile uploadFile = fileStoreService.storeFile(file, isImageFile);
-
         uploadFileDomainService.save(uploadFile);
 
+        // 파일 접근 경로 리턴
         String accessUrlPrefix = fileStoreService.getAccessUrlPrefix();
         String uuid = uploadFile.getUuid();
         String extension = uploadFile.getExtension();
 
-        // 이미지 파일이라면
+        // 이미지 파일이라면 전체 접근 경로 리턴
         if (isImageExtension(uploadFile.getExtension()))
             return accessUrlPrefix + uuid + ensureDotPrefix(extension);
 
@@ -50,14 +51,16 @@ public class FileService {
     // 첨부파일 다운로드 주소 리턴
     public List<AttachedFileInfo> getAttachedFileDownloadPath(long boardId) {
 
+        // 게시글에 연결된 모든 fileId(UploadFile PK)를 리턴
         List<Long> fileIds = boardFileDomainService.findByBoardIdIn(boardId);
+
+        // fileId를 바탕으로 모든 UploadFile을 조회
         List<UploadFile> uploadFiles = uploadFileDomainService.findByIdIn(fileIds);
 
+        // 첨부파일 정보 리턴(파일 이름, 다운로드 경로, 확장자)
         List<AttachedFileInfo> response = new ArrayList<>();
-
-        for (UploadFile file : uploadFiles) {
+        for (UploadFile file : uploadFiles)
             response.add(fileStoreService.getAttachedFileInfo(file));
-        }
 
         return response;
     }
